tests: unify
